Added logging.py, refactored logging in downloaders.py, relocated config files into dataset/hcp/conf (Issue #10)
